+++ test/unfold_settings.py
@@ -2,34 +2,36 @@
 
 from test.utils import get_available_devices, make_id
 
+import torch
+
 DEVICES = get_available_devices()
 DEVICES_ID = [f"device={dev}" for dev in DEVICES]
 
 PROBLEMS_1D = [
     {
         "seed": 0,
-        "input_shape": (2, 3, 50),
+        "input_fn": lambda: torch.rand(2, 3, 50),
         "unfold_kwargs": {
             "kernel_size": 1,
         },
     },
     {
         "seed": 1,
-        "input_shape": (2, 3, 50),
+        "input_fn": lambda: torch.rand(2, 3, 50),
         "unfold_kwargs": {
             "kernel_size": 2,
         },
     },
     {
         "seed": 2,
-        "input_shape": (2, 3, 50),
+        "input_fn": lambda: torch.rand(2, 3, 50),
         "unfold_kwargs": {
             "kernel_size": 3,
         },
     },
     {
         "seed": 3,
-        "input_shape": (2, 3, 50),
+        "input_fn": lambda: torch.rand(2, 3, 50),
         "unfold_kwargs": {
             "kernel_size": 3,
             "dilation": 2,
@@ -37,7 +39,7 @@ PROBLEMS_1D = [
     },
     {
         "seed": 4,
-        "input_shape": (2, 3, 50),
+        "input_fn": lambda: torch.rand(2, 3, 50),
         "unfold_kwargs": {
             "kernel_size": 3,
             "dilation": 2,
@@ -46,7 +48,7 @@ PROBLEMS_1D = [
     },
     {
         "seed": 5,
-        "input_shape": (2, 3, 50),
+        "input_fn": lambda: torch.rand(2, 3, 50),
         "unfold_kwargs": {
             "kernel_size": 3,
             "dilation": 2,
@@ -60,28 +62,28 @@ PROBLEMS_1D_IDS = [make_id(problem) for problem in PROBLEMS_1D]
 PROBLEMS_2D = [
     {
         "seed": 0,
-        "input_shape": (2, 3, 50, 40),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40),
         "unfold_kwargs": {
             "kernel_size": 1,
         },
     },
     {
         "seed": 1,
-        "input_shape": (2, 3, 50, 40),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40),
         "unfold_kwargs": {
             "kernel_size": 2,
         },
     },
     {
         "seed": 2,
-        "input_shape": (2, 3, 50, 40),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40),
         "unfold_kwargs": {
             "kernel_size": (3, 2),
         },
     },
     {
         "seed": 3,
-        "input_shape": (2, 3, 50, 40),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40),
         "unfold_kwargs": {
             "kernel_size": (3, 2),
             "dilation": 2,
@@ -89,7 +91,7 @@ PROBLEMS_2D = [
     },
     {
         "seed": 4,
-        "input_shape": (2, 3, 50, 40),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40),
         "unfold_kwargs": {
             "kernel_size": (3, 2),
             "dilation": 2,
@@ -98,13 +100,21 @@ PROBLEMS_2D = [
     },
     {
         "seed": 5,
-        "input_shape": (2, 3, 50, 40),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40),
         "unfold_kwargs": {
             "kernel_size": (3, 2),
             "dilation": 2,
             "padding": 1,
             "stride": 2,
         },
+    },
+    {
+        "seed": 0,
+        "input_fn": lambda: torch.rand(2, 3, 50, 40, dtype=torch.float64),
+        "unfold_kwargs": {
+            "kernel_size": 1,
+        },
+        "id": "bug-float-64-input",
     },
 ]
 PROBLEMS_2D_IDS = [make_id(problem) for problem in PROBLEMS_2D]
@@ -113,7 +123,7 @@ PROBLEMS_2D_IDS = [make_id(problem) for problem in PROBLEMS_2D]
 PROBLEMS_3D = [
     {
         "seed": 0,
-        "input_shape": (2, 3, 50, 40, 30),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40, 30),
         "unfold_kwargs": {
             "kernel_size": 1,
         },
@@ -121,7 +131,7 @@ PROBLEMS_3D = [
     },
     {
         "seed": 1,
-        "input_shape": (2, 3, 50, 40, 30),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40, 30),
         "unfold_kwargs": {
             "kernel_size": 2,
         },
@@ -129,7 +139,7 @@ PROBLEMS_3D = [
     },
     {
         "seed": 2,
-        "input_shape": (2, 3, 50, 40, 30),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40, 30),
         "unfold_kwargs": {
             "kernel_size": (4, 3, 2),
         },
@@ -137,7 +147,7 @@ PROBLEMS_3D = [
     },
     {
         "seed": 3,
-        "input_shape": (2, 3, 50, 40, 30),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40, 30),
         "unfold_kwargs": {
             "kernel_size": (4, 3, 2),
             "dilation": 2,
@@ -146,7 +156,7 @@ PROBLEMS_3D = [
     },
     {
         "seed": 4,
-        "input_shape": (2, 3, 50, 40, 30),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40, 30),
         "unfold_kwargs": {
             "kernel_size": (4, 3, 2),
             "dilation": 2,
@@ -156,7 +166,7 @@ PROBLEMS_3D = [
     },
     {
         "seed": 5,
-        "input_shape": (2, 3, 50, 40, 30),
+        "input_fn": lambda: torch.rand(2, 3, 50, 40, 30),
         "unfold_kwargs": {
             "kernel_size": (4, 3, 2),
             "dilation": 2,
@@ -167,9 +177,3 @@ PROBLEMS_3D = [
     },
 ]
 PROBLEMS_3D_IDS = [make_id(problem) for problem in PROBLEMS_3D]
-
-UNSUPPORTED_N = [4, -1]
-UNSUPPORTED_N_IDS = [f"N={n}" for n in UNSUPPORTED_N]
-
-UNSUPPORTED_KERNEL_SIZE = [[1, 2], 1]
-UNSUPPORTED_KERNEL_SIZE_IDS = [f"kernel_size={s}" for s in UNSUPPORTED_KERNEL_SIZE]