+++ changelog.md
@@ -7,10 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2021-05-05
+
+### Added
+
+- Generalization of `torch.nn.Fold`
+  [[PR](https://github.com/f-dangel/unfoldNd/pull/18)]
+
+### Fixed
+
+- Bug related to data type of input to `unfoldNd`
+  [[PR](https://github.com/f-dangel/unfoldNd/pull/21)]
+
+### Internal
+
+- Move to `setup.cfg`, automate PyPI release
+  [[PR](https://github.com/f-dangel/unfoldNd/pull/22)]
+- Import sorting
+  [[PR](https://github.com/f-dangel/unfoldNd/pull/19)]
+
 ## [0.0.1] - 2021-02-19
 
 Initial release
 
-[Unreleased]: https://github.com/f-dangel/unfoldNd/compare/0.0.1...HEAD
-<!-- [0.0.2]: https://github.com/f-dangel/unfoldNd/compare/0.0.1...0.0.2 -->
+[Unreleased]: https://github.com/f-dangel/unfoldNd/compare/0.1.0...HEAD
+[0.1.0]: https://github.com/f-dangel/unfoldNd/compare/0.0.1...0.1.0
 [0.0.1]: https://github.com/f-dangel/unfoldNd/releases/tag/0.0.1